LPS-199599 Inline SpeedwellLayoutsInitializer methods into SpeedwellSiteInitializer
