+++ modules/apps/commerce/commerce-theme-speedwell/commerce-theme-speedwell-site-initializer/src/main/java/com/liferay/commerce/theme/speedwell/site/initializer/internal/SpeedwellSiteInitializer.java
@@ -182,7 +182,7 @@ public class SpeedwellSiteInitializer implements SiteInitializer {
 
 			_configureB2CSite(commerceChannel.getGroup(), serviceContext);
 
-			_speedwellLayoutsInitializer.initialize(serviceContext);
+			_initialize(serviceContext);
 
 			_importAssetCategories(serviceContext);
 
@@ -349,6 +349,18 @@ public class SpeedwellSiteInitializer implements SiteInitializer {
 			group.getName(serviceContext.getLanguageId()) + " Portal",
 			CommerceChannelConstants.CHANNEL_TYPE_SITE, null,
 			commerceCatalog.getCommerceCurrencyCode(), serviceContext);
+	}
+
+	private void _createLayouts(ServiceContext serviceContext)
+		throws Exception {
+
+		JSONArray jsonArray = _jsonFactory.createJSONArray(
+			SpeedwellDependencyResolverUtil.getJSON("layouts.json"));
+
+		_cpFileImporter.createLayouts(
+			jsonArray, SpeedwellDependencyResolverUtil.getImageClassLoader(),
+			SpeedwellDependencyResolverUtil.getImageDependencyPath(),
+			serviceContext);
 	}
 
 	private void _createRoles(ServiceContext serviceContext) throws Exception {
@@ -845,6 +857,12 @@ public class SpeedwellSiteInitializer implements SiteInitializer {
 			serviceContext.getUserId());
 	}
 
+	private void _initialize(ServiceContext serviceContext) throws Exception {
+		_cpFileImporter.cleanLayouts(serviceContext);
+
+		_createLayouts(serviceContext);
+	}
+
 	private void _setCommerceShippingMethod(
 			long groupId, String shippingMethod, ServiceContext serviceContext)
 		throws PortalException {
@@ -1123,9 +1141,6 @@ public class SpeedwellSiteInitializer implements SiteInitializer {
 		target = "(osgi.web.symbolicname=com.liferay.commerce.theme.speedwell.site.initializer)"
 	)
 	private ServletContext _servletContext;
-
-	@Reference
-	private SpeedwellLayoutsInitializer _speedwellLayoutsInitializer;
 
 	@Reference
 	private ThemeLocalService _themeLocalService;
